LayoutTests/accessibility/ios-simulator/fieldset-traits.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script>
var successfullyParsed = false;
</script>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body id="body">

<form>
<fieldset id="fieldset">
<legend>Personalia:</legend>
Name: <input type="text" id="inputname"><br>
Email: <input type="text"><br>
</fieldset>
</form>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that elements in a fieldset behave as expected.");

    if (window.accessibilityController) {

        var inputname = accessibilityController.accessibleElementById("inputname");
        var fieldset = accessibilityController.accessibleElementById("fieldset");
        shouldBeTrue("fieldset.isEqual(inputname.fieldsetAncestorElement())");
    }

    successfullyParsed = true;
</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
