LayoutTests/accessibility/ios-simulator/fieldset-traits-expected.txt

Personalia:
Name:
Email:
This tests that elements in a fieldset behave as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS fieldset.isEqual(inputname.fieldsetAncestorElement()) is true
PASS successfullyParsed is true

TEST COMPLETE
